Edit process view in maximized form
