+++ src/plugins/spark/features/ProcessView/grid.styl
@@ -1,0 +1,33 @@
+@import '../../../../../src/styles/quasar.styl';
+
+.grid-base {
+  width: 100%;
+  height: 100%;
+  background-size: 50px 50px, 50px 50px;
+  background-position: 0 -1px, -1px 0;
+  border: 1px solid $dark_bright;
+}
+
+.grid-editable {
+  background-image: linear-gradient(
+    $dark_bright 1px,
+    transparent 1px
+  ), linear-gradient(
+    90deg,
+    $dark_bright 1px,
+    transparent 1px
+  );
+}
+
+.grid-item {
+  position: relative;
+}
+
+.grid-item-coordinates {
+  font-size: x-small;
+  z-index: 2;
+  -webkit-user-select: none;
+  -moz-user-select: none;
+  -ms-user-select: none;
+  user-select: none;
+}
